Fix that secret chats were not opened from the contact list

Closes #142

See merge request ubports/apps/teleports!252

+++ libs/qtdlib/chat/qtdchatlistmodel.h
@@ -40,7 +40,7 @@ public:
 
 public slots:
     void setCurrentChat(QTdChat *currentChat);
-    void setCurrentChatById(const int &chatId);
+    void setCurrentChatById(const qint64 &chatId);
     void createOrOpenPrivateChat(const int &userId);
     void createOrOpenSavedMessages();
     void createOrOpenSecretChat(const int &userId);
